Add useEffect into overview to watch for styles changing on
