+++ client/src/components/overview/FavoriteButton.jsx
@@ -1,9 +1,9 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { regular, solid } from '@fortawesome/fontawesome-svg-core/import.macro';
 
-const FavoriteButtonContainer = styled.div`
+const FavoriteButtonContainer = styled.button`
   position: absolute;
   bottom: 2.5%;
   left: 50%;
@@ -17,18 +17,34 @@ const FavoriteButtonContainer = styled.div`
   backdrop-filter: blur(2);
   transition: all .2s ease;
   box-shadow: 1px 1px 3px rgba(0,0,0,.25);
+  border: none;
 
   &:hover {
     cursor: pointer;
     background-color: rgb(246,246,246);
   }
+
+  &:focus {
+    outline: none;
+  }
+
+  &[disabled] {
+    cursor: not-allowed;
+  }
 `;
 
-const FavoriteButton = () => {
+const FavoriteButton = ({currentStyle, disabled}) => {
   const [favorited, setFavorite] = useState(false);
 
+  useEffect(() => {
+    setFavorite(false);
+  }, [currentStyle]);
+
   return (
-    <FavoriteButtonContainer onClick={() => setFavorite((prevState) => !prevState)}>
+    <FavoriteButtonContainer
+      disabled={disabled}
+      onClick={() => setFavorite((prevState) => !prevState)}
+    >
       {favorited
         ? <FontAwesomeIcon icon={solid('heart')} />
         : <FontAwesomeIcon icon={regular('heart')} /> }
