+++ client/src/components/styled/QuantityDropDown.styled.js
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 const QuantityDropDown = styled.select`
   border: 1px solid #555555;
-  color: #555555;
+  color: black;
   padding: .45em 1.25em;
   background-color: #f3f3f3;
   font-weight: bold;
